// src/ts/storage/enum.ts
export enum CAMPAIGN_STATUS {
	NOT_APPROVED,
	APPROVED,
	OFFERED,
	WAITING,
	ACTIVE,
	END,
	CANCELED
}

export enum COLLECTION {
	ACTIVITY = "activities",
	CAMPAIGNS = "campaigns",
	CONSTANTS = "constants",
	CONTENT = "content",
	DEFAULT_CONTENT = "defaultContent",
	GROUPS = "groups",
	OFFERS = "offers",
	PLAYBACKS = "playbacks",
	PLAYLISTS = "playlists",
	RATES = "rates",
	ROLES = "roles",
	USERS = "users",
	ZONES = "zones"
}